Moving single use Opportunity model to IndexViewModel
